Issue Number: 6380634 & 6417938 Tests Run: PE QL, sqetests/devtests/security/wss/ssl, devtests/registry Submitted By: Harpreet Singh Reviewed By: Satish Vishwanatham
1. 6380634
Unpublish did not remove ExternalLink, Service and ServiceBinding. Previous code was written under the assumption that once and Organization was deleted, all its dependencies would be deleted - this is not the case. Fix was to explicitly go in and delete all dependencies before an Organization is deleted.

2. 6417938
A SSl webservice's wsdl was not published correctly. The regex code that looked through the WSDL to determine host/port information was only looking up the string "http". It has been modified to look at https as well.