Issue number:  691 Obtained from: Issue Tracker Tested by:  QL, entity-persistence-tests against Oracle Reviewed by:   Tom Ware
[NEWS:entity-persitence 691] IN=99212 - Generate BLOB or CLOB columns for Lob types on Oracle

The problem is java2db creating a table having two LONG fields on Oracle. Oracle
does not permit to create such a table, and the test fails. The entity in the
test has two fields that are annotated @Lob,

  @Lob
  @Column(name = "BLOBDATA")
  public byte[] getBlobData(){

and

  @Lob
  @Column(name = "CLOBDATA")
  public String getClobData(){

Because of the @Lob annotation, these fields are recognized by TopLink as type
java.sql.Blob and java.sql.Clob respectively. I've changed OraclePlatform to
register column type BLOB and CLOB the these java types in fieldTypeMapping.
Previously the column types where LONG RAW and LONG. After the change, the table
has a BLOB and a CLOB column, and can be created successfully.