Issue Number:6468489 Submitted By: Harpreet Singh Reviewed By: Self (This is a minor variation on 6380634 which was reviewed by Satish) Tests Run: PE QL, 3 devtests/registry against UDDI and EBXML registry.
For bug id 6380634 - I made a fix to unpublish ExternalLink that were published to a registry. As part of clean up of this code, I added ExternalLink to service in addition to an Organization object (earlier a ExternalLink was attached to just an Organization object).
UDDI does not permit ExternalLinks to Services. Hence the publish to UDDI bombs. The fix is fairly simple - to check if publish operation is to a UDDI registry and attach ExternalLink to the Organization.

Files Modified: RegistryAccessObjectImpl