[NEWS:entity-persistence 702] IN=100458 - Add validation preventing primary key updates
The pk update validation is implemented in
writeFromObjectIntoRowWithChangeRecord() of classes AbstractDirectMapping and
OneToOneMapping. I added a ValidationException and changed
ValidationExceptionResource for that purpose. EntityManagerJUnitTestSuite now
has three new tests:

a) that the correct exception if thrown in case of a primary key update
b) that updating a primary key field with the same value is allowed
c) that the correct exception if thrown in case of a primary key update in an
overlapping primary/foreign key situation

Issue number: 702
Obtained from: IssueTracker
Tested by: QL, entity-persistence-tests
Reviewed by: Gordon Yorke