Issue number:  2139706 Obtained from: - Submitted by:  Sivakumar Thyagarajan Reviewed by: Forward porting of bug fix for 6430394
Changed files:
appserv-core/src/java//com/sun/enterprise/resource/ResourceManagerImpl.java

Description of fix: at certain points, when a XA resource could not be
registered as part of a component's resource list, due to a failure
while enlisting the XAResource in the transaction, we weren't cleaning
up the resource from the component list. This resulted in the same
XAResource being reused when the component(MDB) is reused. The current
fix ensures that the resource is cleaned from the component list when
there is failure elisting the XA resource in the transaction

Tests run : PE QL, Connector dev tests,