Issue number: 694 Reviewed by: Kedar Mhaswade, Byron Nevin
Modified KeystoreManager.java to use ProcessBuilder(List<String>) instead of Runtime.exec(String).  This will f
ix the issue with space in the directory path.
Added code in KeystoreManager.java to check if the args are non-null and that file exists.
Added entry in LocalStrings.properties for "fileNotFound".
In all pe80/startserv.tomcat.template, pe80/stopserv.tomcat.template, ee80/startserv.tomcat.template and ee80/s
topserv/tomcat.template added "quote" for ASENV_CONF_LOCATION and INSTANCE_ROOT so that it's possible to have s
pace in the directory location.

Manually tested on both Unix and Mac with a space in the directory path.
Passed QL tests.