Fix for 6475894 - undeploy fails when Symbolic links are used
Problem: Undeployment or redeployment fails because the directory deletion logic does not delete files that are referred to via sym links.

Diagnosis: The logic purposefully compares each file's absolute path to its canonical path.  These values are different if the file path contains a sym link.  This logic is in place to avoid following a link from an app server directory to a user directory and deleting the contents of the user directory.  But the logic also incorrectly prevents deletion of directories that the app server properly owns that happen to be referred to using a symlink "upstream" from the app server directory.

Fix: In the FileUtils.whack method, resolve any sym links present in the path to the the File that is passed in to be deleted.  Then as the subdirectories are traversed compare the absolute path to the canonical path.  If they disagree, delete the file which deletes the link but not its contents.  If they agree, delete a as before, recursively if it is as directory.


Reviewed by: Byron


Tests:

PE SQE quick-looks; deployment devtests


Code changes:

appserv-commons/src/java/com/sun/enterprise/util/io/FileUtils.java