Issue number: 6454697 Obtained from: Submitted by:  Kshitiz Saxena Reviewed by:   Siraj Gaffar
The JMS/Connector service monitoring was not getting enabled dynamically. There was issue with condition of app name being non-null, even for system as well as standalone RAs. The case of app name being null needs to be handled for system and standalone RA.

Tests run: QL, jdbc/connector devtests