Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=1301 ("Change semantics of <http-listener> "enabled" attribute to truely reflect an HTTP listener's state")
Tests run: QL

[NEWS:webcontainer 1301] Change semantics of <http-listener> "enabled"
attribute to truely reflect an HTTP listener's state
Changed the semantics of the <http-listener> "enabled" attribute, so
that if set to FALSE, the http listener will not be started, or if the
http listener was already started, it will be destroyed and will
release its port. By subsequently setting "enabled" to TRUE, the http
listener will be (re)started. Changes to the <http-listener> "enabled"
attribute are reflected immediately, that is, without requiring a
server restart.