[NEWS:entity-persistence 1021] IN=102035 - Register new non-entity instances in identity map of closest entity superclass
With this integration, new instances of non-entity subclasses will be
registered in the identity map of the closest entity superclass. This
is achieved by changing the classType attribute of OjectChangeSet from
the actual classType to the class of the closest entity
superclass.

Issue number: 1021
Obtained from: IssueTracker
Tested by: QL, entity-persistence-tests
Reviewed by: Gordon Yorke