Fix Issue 1314: https://glassfish.dev.java.net/issues/show_bug.cgi?id=1314 ("Regression: javax.servlet.request.X509Certificate not added to the session")
When The client sent the finished message, and then immediately sent the HTTP headers without doing any further reads, the headers will be sent under the non-auth session. Even if the server decides right after sending its FINISHED message that it wants to switch to client auth, the client can still send as much application data (HTTP headers) as it wants before actually seeing and responding to the ClientHelloRequest message. In general HTTP traffic, it will only notice the ClientHelloRequest when the client wants the server's response to the HTTP message. Hence we need to make sure a re-handshake happens every time the certificates are requested by the client. With NIO non-blocking, we can't assume the certificate will be available until the seconds handshake happens.

The JSSE implementation will obtain a list of accepted cert issuers from the X509TrustManager, and will send that list as part of the CertificateRequest message.  The client is supposed to consult its list of credentials, and send a credential which was issued by one of the cert issuers found in the received CertificateRequest message.  If the client is not doing so (asadmin), then it's possible to not have the client credentials set up correctly. To cover that case, the setWantClientAuth(true) must be called instead of setNeedClientAuth(true), because setting the later will make the handshake fail if the client ignore the request.

- ALL QL passed
- All devtests/web passed (including the web/servletSSLRequestAttributes)
- ALL Servlet TCKs passed
- ALL JSP tcks passed
- All devtests/security/cipher passed
- All devtests/security/web-mutual-auth

Reviewed by: Jan Luehe
Kudo to Brad Wetmore (SSL/NIO lead) for the help on this