The following changes were made to the admin console jbi screens.
a. Made changes to the jsf files and added some javascript so some of the table components would be disabled or enabled based on a row being selected in the table.

b. Fix the code in the sharedTable.jsf file so the check all and uncheck all buttons would work.

c. Added the code to the main page jsf files so the context sensitive help would work correctly.

Reviewed-by: Mike Wright
Tests-run: manual tests, "ant jbi-test", quicklook tests (all tests passed)