[NEWS: cmp 1400]
This integration is the runtime part of a performance improvement to
be implemented in CMP code-gen. The default fetch-group fields in the
JDOState classes generated for CMP beans used to be updatable w/o
StateManager interaction.  Because of that, all DFG fields are marked
as updated, if any of the DFG fields is updated. In the future, the
code-gen will be changed to allow individual change tracking for DFG
fields. In JDO words, the enhancement schema for DFG fields will be
changed from CHECK-READ/CHECK-WRITE to CHECK-READ/MEDIATE
WRITE.Therefore the changed DFG fields have to be compared for
batching compatibility, which wasn't done before based on the
assumption, that all DFG fields are written back in case of updates.

Issue number: 1400
Obtained from: IssueTracker
Tested by: QL, cmp-tests (runtime_all.list), CTS
Reviewed by: Mitesh Meswani