Issue number: 726, bug# 6179968 Reviewed by:  Prashanth Abbagani
This is to fix the issue with returning the exit code if domain/nodeagent/cluster/instance
is already running or stopped.  Currently if you try to start a domain that has already started,
the exit code is 1.  The fix is to return the exit code of 0 if domain has already started.
Likewise for the stop-domain command.  This applies to nodeagent, cluster and instance as well.

passed QL tests
passed manual tests