Issue number: 6378839 Reviewed by:  Kedar Mhaswade
Add a portbase option in create-domain command.
The logic to this portbase is as follow:
(1) if portbase is specified and either adminport, instanceport or domainproperites options is also specified then fai
l the command stating that the options are mutually exclusive.
(2) if any of the ports are below 1024, then display a warning message.
(3) portbase is added to the following ports:

   Admin port  = portbase + 48
   Instance port = portbase + 80
   JMS port = portbase + 76
   IIOP port = portbase + 37
   HTTPS port = portbase + 81
   IIOPS port = portbase + 38
   IIOP MutualAuth port = portbase + 39
   JMX port  = portbase + 86

passed QL tests
passed Manual tests