Submitted by:  Self Reviewed by:   Dinesh Patil.
This checkin hopefully modifies the right templates.
Only templates are modified for necessary properties
with this checkin.

This is just the background work.
This checkin means that if the tokens exist,
they will be replaced. No side effects.
Code is not fully integrated yet. None is reading
these values as of this checkin.