Partial fix for improving deployment performance plus fix for 6175691 - application re-deploy to domain changes the application-ref
In the past redeployment would remove all app refs for the app being redeployed, then redeploy the app to the DAS, then add app refs again. The removal and readdition of app refs is costly because they trigger apparent config changes that must be saved, even if though the new values are the same as the old ones.

Also, the logic did not preserve any non-default values on app refs that are not settable via the deployment properties. For example, the lbEnabled value can be set only using dotted notation (not via options on the deploy command). If a user did so and then redeployed the app, the resulting app ref would lose the lbEnabled setting that the user had assigned with dotted notation.

This set of changes allows the back-end to detect, during all phases, if a redeployment is in progress. If so, and if an app ref is not being changed by any deployment-time options, then the app ref is NOT deleted by disassociate or added back by the later associate. Further, if the delete and add-back is needed, the original app ref DOL object is saved temporarily during disassociate. Then during associate it is retrieved and after it is added back the values settable only via dotted notation are assigned to the newly-added-back app ref. This preserves the user's manual settings of such attributes.

Synopsis of changes below with list of changed files

Reviewed by: Hong


Tests:

PE and EE SQE quick-looks; deployment PE and EE devtests; manual tests to verify that non-default attribute settings of app refs are preserved across redeployment


Code changes:
===== appserv-commons/src/java/com/sun/enterprise/deployment/util/
DeploymentProperties.java - REDEPLOY property should not be subject to AMX conversion; AMX does not support it.

===== appserv-core/src/java/com/sun/enterprise/deployment/backend/
DeploymentRequest.java - Add isRedeployInProgress and setter methods and the corresponding field.

===== appserv-core/src/java/com/sun/enterprise/deployment/phasing/
ApplicationReferenceHelper.java - Add findCurrentAppRef method to locate the app ref object for a given app and target.

AssociationPhase.java - Recreate the app ref only if an attr value settable via deployment has changed; also when recreating an app ref during redeployment get values not settable on the deploy command from the information about the app ref saved during disassociation.

DeploymentContext.java - Add inner class that records information about an app ref; add save, get, and remove methods and supporting HashMaps.

DeploymentServiceUtils.java - Reuse existing getDeploymentTarget method instead of duplicating the same code.

DisassociationPhase.java - Do not remove an app ref during redeployment if its attr values are not changing.

PEDeploymentService.java - During redeployment do not validate that a ref about to be added is not already there if the app ref was not actually removed during the disassociate phase; also set the isRedeploymentInProgress in the deployment options for all phases.

===== appserv-core/src/java/com/sun/enterprise/deployment/client/
DeployAction.java - Optimize the isModuleDeployed call to the back-end; also set the REDEPLOY deployment property for the phases.