Issue #:1629
When showing detail information about an entry in the list (table), the entry was not
being highlighted. Highlighted in this case is bolding the name column and displaying
the small select bullet icon in front of the name.  The name filed will also be switched
from a hyperlink to a static text entry. This change was implemented in both PE and EE
screens.

As a bonus, this integration also included the addition of the verification deletion
dialog box.  This confirmation dialog will appear when any deletions is being performed
from one of the tables.

Reviewed By: Mike Write

Tests-run: QL, jbi-test