Issue number:  1558 Submitted by:  Byron Nevins Reviewed by:   Nandini Ektare The penultimate cause of this bug:
PE used to always have changeit as the master password.  The original code
detected PE by the absence of username/admin password.  In such a case no
security info was sent to the server, since the master password was guaranteed
to be changeit, and the appserver knew about changeit.  Thus, I could always get
away with sending no security info to the server in PE.

Now developer-domains can have a user-chosen master password.  The original code
would never send in the master password -- which caused the bug.

The fix is to always send in the security info -- even if username/admin
password are null.