Fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=1699 ("LinkedListPipeline thread count is wrong")
LinkedListPipeline.addTask() gets the current queue size to determine if a task should be rejected and then reuses that size when determining if a new thread should be started. It should use queueSize + 1 to determine if the new thread should be started (since it is added a task)

Patch provided by Scott Oaks.

- All QL passed.