Issue #:1694
The dropdown filters for both the Component table and the Service Assembly table
were not filtering the table on the state field correctly.  The EE screens were
changed to use "Enabled" and "Disabled" instead of "Started" and "Stopped".

A dropdown filter was added to the Target Table using the Status column.

Reviewed By: Mike Write

Tests-run: QL, jbi-test