Issue #:1728
The code that I sent you actually displays the state of the component, appended to the end of the string.  This was helpful in developing the GUI.  This will be removed shortly, since it would not be relevant unless a specific target was specified.

The filter option on the lifecycle state for the jbi service assemblies, binding
components and service engines was not working correctly.  The code was fixed to
solve this problem.

updated JBICommands.java
updated JBIListCommands.java

Reviewed By: Jane Young