Issue number:  1802 Synopsis:      Several improvements proposal for entity-persistence-tests. Submitted by:  Wonseok Kim Reviewed by:   Tom Ware
* Run test with another DB properties without rebuilding
Database properties are separated from build.properties and stored in test.properties file.
Modify the test.properties file directly or use another properties file with -Dtest.properties option for ant.

To make this possible the properties file is read from testing framework(JUnitTestCase) in runtime and they override the properties which is passed to creatEntityManagerFactory(pu, properties).
So we don't need to put these properties in persistence.xml files and replace it in build time, it is done now in runtime. "ant build" is not any more dependent on properties file, so it doesn't need to rebuild for another database properties.

I put derby database properties in test.properties for default configuration.
This makes beginners quick to run test out of the box with bundled JavaDB.

* Run a single test class
To run a single test class, -Dtest.class option can be used.

* Log level without modifying persistence.xml files
test.properties file can contain toplink.logging.level property and this will override the persistence.xml property.

The junit ant task in build.xml now contains maxmemory="256m" because it threw OutOfMemoryError if log level is FINEST.