Issue number: bugster: 6505272
revert back changes for start/stop-cluster commands:
CLIDescriptor.xml: rev 1.5 : start/stop-cluster commands returns to invocation of GenericCommand.java
J2EELifecycleCommand.java: rev 1.1 :  remove checks for start/stop-cluster commands
StartClusterCommand.java: rev 1.1 : to extend from GenericCommand instead of J2EELifecycleCommand

Agreed to revert back the changes for start/stop-cluster commands.
By reverting this change, the start/stop-cluster commands will not return an exit code of 0 if the cluster is already started/stopped.  This is an acceptable behavior since cluster is not a running instance.  Cluster is a group of homogeneous instances.  If one server is not running while all the servers in the cluster are running, the cluster is considered a partial-running state,  However, this state is not respected by JSR77 thus the backend implementation has no such state.  Therefore a parital running state is considered as a running state.
Revertying the change will cause the start-cluster command to  always start all the server instances.  The caveat to this is that if all the instances in the cluster are running and start-cluster is executed then the command will fail since all instances are already running.