Issue #:1965
Fixed the problem where the alert messages were not being display when there
was an error while managing targets.  Also added a fix to the code that will
display an alert when the user tries to perform a lifecycle operation to a
component or a service assembly when it is not associated to a target.

Reviewed By: Mike Write

Tests-run: QL, jbi-test