Issue number:   1615 Obtained from:  IssueTracker Submitted by:   Vince Reviewed by:    Self
This is the last step in this fix.
Removed the values from asenv template as they
will be taken from asadminenv.conf.
Now, all the defaults for asadmin will be
available. I have tested in such a way that
QL does NOT pass the values explicitly and
this tests if the defaults are appropriately
handled by asadmin or not. I am going to file
a separate RFE for that against QL testsuite.