Issue #:1880
The enable button will now start the component / service assembly after it
has been installed or deployed.

Alert message will be display when the selects the Enable checkbox and
the component was not installed to any target.

If the user tries to enable the component via the Enable button and the component
is not associated with a target, an alert will now be displayed

Also fixed a bug, deletion would not work if you tried to delete
a component or a service assemble without first going to the show screen.

Reviewed By: Mike Write

Tests-run: QL, jbi-test