Issue number:   6512455 Obtained from:  Sun's Bug Database Submitted by:   Nitya Reviewed by:    Self
Removed the complex and hardcoded logic for
the supported cipher suites. GUI makes it
impossible to choose unsupported cipher suites
anyway.

Will work on this again, in order to fix it
completely. For now, GUI should have no problems
saving the cipher suites.

QL passed.

A comment is changed in the DTD. There is
no other change.