Issue number:  Sun's bug database: 6510947 Obtained from: Upgrade Submitted by:  Upgrade Reviewed by:   Self Approved by BugSwat.
This is a DTD change. We are making the
GMS attributes optional on cluster element.
This is taken care of by the Java code in
that a cluster element "requires" these
elements, but DTD does not mandate it.

This is needed for upgrade scenarios.

It was a mistake to make these required
in sun-domain_1_2.dtd. Correcting the same.

developer/cluster QL passed.
Manual testing with create clusters succeeded.