Incremental fix for https://glassfish.dev.java.net/issues/show_bug.cgi?id=2103 ("FORM-based auth failing with in-memory-replication enabled"):
When calling request.isRequestedSessionIdValid(), check if the request
has already cached a session whose id matches the requested id, and if so,
return session.isValid(), instead of calling manager.findSession() with the
requested id.
Calling manager.findSession() must be avoided in the case where the
manager supports session versioning, because there might be a mismatch
between the requested session version and the incremented session version
of the manager's cached session (there will be a mismatch if
request.getSession() was called before request.isRequestedSessionIdValid()).

Reviewed by: lwhite

Tests run: QL, Servlet/TCK