Issue number:   2300 Obtained from:  Issue Tracker Submitted by:   Self Reviewed by:    Self
The custom profile support is added.
Now, one can provide a set of style sheets to transform
the domain.xml in a particular way. This gives the user
the flexibility to create domains with custom usage profile.

The support is only available for domain.xml at the moment.
For GlassFish V2, we will keep it for domain.xml alone.

Ran QL tests against developer profile and cluster profile.
The code is changed in such a way that the usual domain
creation is not modified at all. Only if domain.xml template
does NOT exist at a particular profile folder (developer, cluster
and enterprise) will the support kick in. And in our
distribution, we always provide the domain.xml template.

Manual testing shows that the existing template and using the
transforms works equally well.

I am going to keep this bug open so that:
- it could be documented, and
- I could work on providing an example profile.