Issue Number: 6497618 Submitted By: Harpreet Singh Reviewed By: Satish Vishwanatham Tests Run: PE QL, devtests/registry
Web Services backend modified code to generate WSDL on the fly and not have an in memory representation of this particular file. This WSDL was required to publish the code to
the registry and for extracting the port information on the WSDL. The WebServicesMgrBackend used to store this WSDL as part of a HashMap, WS registry code used to just look it
up.
The code has been modified to a) extract port information by making a MBean request to http-listener of the DAS. b) Extract the URI from WebServiceEndpointInfo and construct a
URL to make a request to the URL to get the WSDL back. This WSDL is then used as it was used earlier.
Note: I am downgrading this as a P3 - as WebServiceEndpointInfo.getWSDL should return the WSDL for the particular web service. This used to work earlier *I think* but the back
end code change has bitten this as well. Once this is changed - I can change the registry code to use AMX instead of the code I added for this fix.

Files changes: admin/wsmgt/.../Registry*.java, ConfigHelper.java and WebServiceMgrBackend.java.