Issue number:  6521606 Obtained from: Shing Wai (Bugster) Submitted by:  Shing Wai Reviewed by:   Self.
Passed the required property to the JDK keytool,
*only while creating the self-signed cert*, so
that it can be created with skid extension, when
it is available in the keytool.

Tested creation/startup/shutdown of servers.
At the moment, I don't think I have a JDK which
respects this -J-D on the keytool command line.

Requires verification.