Issue number:   2102 Obtained from:  Issue Tracker Submitted by:   Shreedhar Ganapathy Reviewed by:    Self
This was an oversight and thanks for catching it.

Now, if the file .asadminpass is tampered, a more
information message is displayed:

Login information relevant to admin user name [admin] for this domain [dd] could not be stored.
This means that the file .asadminpass in home folder is tampered.
Fix the format for this file.

Tested manually.