IN=106808 - JPQL: fix issue with navigation through two level of embeddeds
Fix problem with an JPQL query navigating two levels of embedded instance in
the SELECT clause. E.g. "SELECT s.company.contact.firstName FROM Submission s",
where contact is embedded into company and company is embedded into submission.
The query compiler sets an outer join flag for this query and this results in
an exception that an outer join is invalid for the embedded fields.

The query compiler needs to set the outer join flag in case a relationship field
is selected in the SELECT clause. The current implementation checks whether a
selected field is not mapped as DirectToFieldMapping. I changed this to directly
check whether the field is mapped as relationship.

Issue number: 2462
Reviewed by: Tom Ware
Tests : QL, entity-persistence-tests, local tests