IN=106892 - JPQL: fix SQLException thrown for constructor query using COUNT
Fix issue with generated SQL for a JPQL query using a constructor expression
with an COUNT argument counting a a joined variable, e.g.
  SELECT NEW test.pojo.CustomerDetail(c.name, COUNT(o))
  FROM Customer c JOIN c.orders o GROUP BY c.name
The generated SQL misses a COUNT argument:
  SELECT t0.NAME, COUNT() FROM CUSTOMER t0, ORDER_TABLE t1 ...

Method ReportQuery.prepareObjectAttributeCount iterates the SELECT expressions
of a ReportQuery in order to execute some special handling for COUNT aggregates
in the SELECT clause. I extended the method to iterate the constructor arguments
in case of a constructor expression. This makes sure we generate the same SQL
whether there is a constructor expression or not.

Issue number: 2496
Reviewed by: Tom Ware
Tests : QL, entity-persistence-tests, local tests