Issue number:   (Bugster) 6509158 Obtained from:  Bugster Submitted by:   Satish Reviewed by:    Satish, Jane (Thanks!)
Introduced a new file called
".upgradedTo<Major><Minor> to optimize the overhead
because of parsing the domain.xml all the time.

This improves startup performance.
The file that is touched and is present in <config> folder
of the domain to be started indicates whether upgrade was
done or not. For any 9.1 domain, e.g. after starting for the first
time, we will have a file called ".upgradedTo91" in its config folder.

This file will be documented.

Ran the QL tests successfully. Manual testing was performed.