Issue number:  2517 Obtained from: Submitted by:  Byron Nevins Reviewed by:   Sreenivas Munnangi
Background:
The Windows Service routine in our code would do the "start" command.  Normally this is "asadmin start-domain ....".
It would then wait for the asadmin process to die and report back that the domain started.
Thread dumps can not appear in server.log because the server's parent process (asadmin) is dead.

Changes:
The service handler now looks for the string "--verbose" in the start command.  If it sees it -- it does not
wait for the asadmin process to die (asadmin won't die until the server is stopped in verbose mode).
The "asadmin start-domain --verbose ...." process is started and Windows immediately says that it has started.
We *could* get the real status and wait for it to be really started before telling Windows.  There is an RFE
that will fix this and other things (6468691)

Summary:
If you create the service with "--verbose" in the asadmin start-domain command, then you WILL see all
threaddumps in server.log