IN=107023 - JPQL: fix NPE for COUNT on variable joining a ManyToMany self-ref relationship
Fix NPE in the toplink expression framework when generating SQL for a query
having a COUNT on a variable that joins a ManyToMany self referencing relationship.
The old code in SQLSelectStatement.appendFromClauseForOuterJoin could not distinguish
source and target table, because they are the same for a ManyToMany self referencing
relationship. I changed this to make use of the table aliases, since they are
different for source and target table.

Issue number: 2440
Reviewed by: Tom Ware
Tests : entity-persistence-tests, local tests