Fix for issues 1428 and 2547. It also fixes 6425413.
Basically outbound sockets in the appserver VM were wrapped
to NIO by quickstartup code. NIO sockets doesnt support
timeouts directly. The recommended practice is to use a
selector to do the timeout. However that is not effective
and cause performance regressions.

The fix is, for non-NIO outbound sockets, wrap plain socket impl
itself, instead of NIO sockets

Reviewed by: Sankar and partly by Scott.
Tests : QL and devtests on Solaris, Windows and Linux.