First Checkpoint integration for the Cluster Logging page.
This Integration includes the cluster logging page jsf file and updates to
the OperationHandlers.java file.  The code in the OperationHandlers.java file
was changed to handle both the PE logging page as well as the EE logging page.
This checkpoint integration will handle the happy path and does not specify any
success alert message after changed the log levels.  The error and successful
alert messages will be implemented in an upcoming integration.

Reviewed By: Mike Write

Tests-run: QL, jbi-test