Fixes for   Issue 848 - ACC starts application clients on the EDT   Issue 2297 - Rich client graceful handling of UnsupportedClassVersionError
Problems:

(848) During a Java Web Start launch, the ACC (and therefore the user's code) starts on the Swing event dispatcher thread rather than the main thread, as it does during appclient script launches.  Although this caused no "errors" it did mean that app clients ran on different threads depending on how they were launched.  Some users objected that they had to write code differently if they had work that should (or should not) run on the event dispatcher thread.

(2297) During an appclient script launch, any errors detected and logged by the ACC itself before user code gains control is written to System.err and, therefore, the command output window.  But during a Java Web Start launch there is no such window.  As a result, errors that occur before the ACC transfers control to the developer's app client - or even errors that occur after that if the app client does not catch and display errors - would be lost.  The end user would see the Java Web Start download and security information but then the client would seem to never run.

Fixes:

(848) Start the JWSACCMain on the main thread, not the event dispatcher thread.  This also requires the ACC main class to do the client log-in on the event dispatcher thread so that information is available when GUI gestures trigger access to protected resources on the back-end.

(2297) In the catch-all handler in the Java Web Start ACC log the exception and stack trace to System.err and also display it in a dialog box that is visible to the end-user.

Reviewed by: Hong


Tests:

PE SQE quick-looks; deployment devtests; manual tests


Code changes:

appserv-core/
  build.xml

  src/java/com/sun/enterprise/appclient/
    MainWithModuleSupport.java
    AppContainer.java

  src/java/com/sun/enterprise/appclient/jws/
    ClassPathManager.java
    ClassPathManager15.java
    ClassPathManager16.java
    JWSACCMain.java

  src/java/com/sun/enterprise/appclient/jws/boot/
    ErrorDisplayDialog.java (new file)
    LocalStrings.properties (new file)