Issue number:  6465513 Obtained from: - Submitted by:  Kshitiz Saxena Reviewed by:   Sivakumar Thyagarajan, Sreenivas Munnangi
Earlier a resource-ref creation was attempted for security map. This was not needed as security map is a global resource. Also security maps are not deleted individually at the time of undeployment. They are deleted automatically when connector connection pool are deleted.

Tests run: QL, connector devtests and CTS Smoke