IN=107276 - JPQL: fix invalid SQL for COUNT on joined variable
Fix the generated SQL for a JPQL query using a COUNT on a joined variable, e.g.
  SELECT c.name, COUNT(o) FROM Customer c LEFT JOIN c.orders o GROUP BY c.name

The fix distinguishes 4 cases:
(1) The entity class of the joined variable has a single primary key. Then the
SQL should be "COUNT(o.PK)" where PK is the primary key column of the order table.
(2) The entity class of the joined variable has a composite primary key and the
COUNT does not use DISTINCT. Then the SQL can do a COUNT on an arbitrary primary
key column.
(3) The entity class of the joined variable has a composite primary key, the
query uses COUNT(DISTINCT o) and the variable o is defined in an inner join.
Then we use the previous code generating a COUNT(*) with an EXISTS subquery.
(4) The entity class of the joined variable has a composite primary key, the
query uses COUNT(DISTINCT o) and the variable o is defined in an outer join. The
fix does not cover this case and an exception in thrown.

Issue number: 2497
Reviewed by:Chris Delahunt
Tests : entity-persistence-tests, local tests